Added range-scan operation in client API

+++ client/src/main/java/io/streamnative/oxia/client/metrics/Counter.java
@@ -44,7 +44,7 @@ public class Counter {
         add(1);
     }
 
-    public void add(int delta) {
+    public void add(long delta) {
         counter.add(delta, attributes);
     }
 }
